added feeds DB, add DB query, and view all feeds query. Created commands agg, addfeed. feeds

+++ handlers.go
@@ -41,34 +41,80 @@ func handlerRegister(s *state, cmd command) error {
 		return err
 	}
 	s.cfg.SetUser(insertedUser.Name)
-	fmt.Printf("%s was created", insertedUser.Name)
+	fmt.Printf("%s was created\n", insertedUser.Name)
 
 	return nil
 }
 
 func handlerReset(s *state, cmd command) error {
+	if len(cmd.Args) != 0 {
+		return fmt.Errorf("usage %s <name>", cmd.Name)
+	}
 	ctx := context.Background()
 	err := s.db.DeleteUsers(ctx)
 	if err != nil {
 		return fmt.Errorf("couldn't delete users: %w", err)
 	}
-	fmt.Printf("DB users deleted")
+	fmt.Printf("DB users deleted\n")
 	return nil
 }
 
 func handlerUsers(s *state, cmd command) error {
+	if len(cmd.Args) != 0 {
+		return fmt.Errorf("usage: %s <name>", cmd.Name)
+	}
 	ctx := context.Background()
 	users, err := s.db.GetUsers(ctx)
 	if err != nil {
 		return fmt.Errorf("unable to retrive users %w", err)
 	}
-	//currentUser := s.cfg.CurrentUserName
 	for _, user := range users {
 		if user.Name == s.cfg.CurrentUserName {
 			fmt.Printf("%s (current)\n", user.Name)
 		} else {
 			fmt.Printf("%s\n", user.Name)
 		}
+	}
+	return nil
+}
+
+func handlerAgg(s *state, cmd command) error {
+	if len(cmd.Args) != 0 {
+		return fmt.Errorf("usage: %s <name>", cmd.Name)
+	}
+	ctx := context.Background()
+	fmt.Print(fetchFeed(ctx, "https://www.wagslane.dev/index.xml"))
+	return nil
+}
+
+func handlerAddFeed(s *state, cmd command) error {
+	if len(cmd.Args) != 2 {
+		return fmt.Errorf("usage: %s <name>", cmd.Name)
+	}
+	ctx := context.Background()
+	user, err := s.db.GetUser(ctx, s.cfg.CurrentUserName)
+	if err != nil {
+		return err
+	}
+	insertedFeed, err := s.db.CreateFeed(ctx, database.CreateFeedParams{ID: uuid.New(), CreatedAt: time.Now(), UpdatedAt: time.Now(), Name: cmd.Args[0], Url: cmd.Args[1], UserID: user.ID})
+	if err != nil {
+		return nil
+	}
+	fmt.Printf(" %s added to database under user %s\n", insertedFeed.Url, user.Name)
+	return nil
+}
+
+func handlerFeeds(s *state, cmd command) error {
+	if len(cmd.Args) != 0 {
+		return fmt.Errorf("usage: %s <name>", cmd.Name)
+	}
+	ctx := context.Background()
+	feeds, err := s.db.GetFeeds(ctx)
+	if err != nil {
+		return err
+	}
+	for _, feed := range feeds {
+		fmt.Printf("%s\n", feed)
 	}
 	return nil
 }
